+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ### `latest`
 
 - Feat: #13 pass old value and all the current values to `on_event`
+- Feat: #14 `send_animation` method for Telegram bots
 
 ### `0.7`